+++ project/src/actions/types.ts
@@ -5,3 +5,5 @@ export const UPDATE_LIST_TITLE = 'UPDATE_LIST_TITLE';
 export const DELETE_LIST = 'DELETE_LIST';
 
 export const ADD_NOTE = 'ADD_NOTE';
+
+export const DRAG_NOTE = 'DRAG_NOTE';